# Conflicts:
#	src/main/java/org/bookstore/controller/BookController.java

+++ src/main/java/org/bookstore/service/impl/BookServiceImpl.java
@@ -49,6 +49,9 @@ public class BookServiceImpl implements BookService {
 
     @Override
     public void delete(Long id) {
+        if (!bookRepository.existsById(id)) {
+            throw new EntityNotFoundException("Can't find book with id: " + id);
+        }
         bookRepository.deleteById(id);
     }
